Expose matcher operations, consumers can extend expect themselves
