+++ packages/graphql-testing/package.json
@@ -8,7 +8,10 @@
   "typesVersions": {
     "*": {
       "matchers": [
-        "./build/ts/matchers/index.d.ts"
+        "./build/ts/matchers/types.d.ts"
+      ],
+      "matchers-all": [
+        "./build/ts/matchers/types.d.ts"
       ]
     }
   },
@@ -47,16 +50,25 @@
     "index.esnext",
     "matchers.js",
     "matchers.mjs",
-    "matchers.esnext"
+    "matchers.esnext",
+    "matchers-all.js",
+    "matchers-all.mjs",
+    "matchers-all.esnext"
   ],
   "module": "index.mjs",
   "esnext": "index.esnext",
   "exports": {
     "./matchers": {
-      "types": "./build/ts/matchers/index.d.ts",
+      "types": "./build/ts/matchers/types.d.ts",
       "esnext": "./matchers.esnext",
       "import": "./matchers.mjs",
       "require": "./matchers.js"
+    },
+    "./matchers-all": {
+      "types": "./build/ts/matchers/types.d.ts",
+      "esnext": "./matchers-all.esnext",
+      "import": "./matchers-all.mjs",
+      "require": "./matchers-all.js"
     },
     ".": {
       "types": "./build/ts/index.d.ts",
