+++ packages/graphql-testing/src/matchers/all.ts
@@ -1,0 +1,5 @@
+import {toHavePerformedGraphQLOperation} from '.';
+
+expect.extend({
+  toHavePerformedGraphQLOperation,
+});
