+++ packages/graphql-testing/src/matchers/types.d.ts
@@ -1,0 +1,13 @@
+import type {GraphQLOperation} from 'graphql-typed';
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace jest {
+    interface Matchers<R, T = {}> {
+      toHavePerformedGraphQLOperation<Variables>(
+        document: GraphQLOperation<any, Variables, any>,
+        variables?: Partial<Variables>,
+      ): void;
+    }
+  }
+}
