+++ packages/react-testing/src/matchers/types.d.ts
@@ -1,0 +1,30 @@
+import type {ComponentType, Context as ReactContext} from 'react';
+
+import type {PropsFor} from '../types';
+
+type PropsFromNode<T> = T extends Node<infer U> ? U : never;
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace jest {
+    interface Matchers<R, T = {}> {
+      toHaveReactProps(props: Partial<PropsFromNode<T>>): void;
+      toHaveReactDataProps(data: {[key: string]: string}): void;
+      toContainReactComponent<Type extends string | ComponentType<any>>(
+        type: Type,
+        props?: Partial<PropsFor<Type>>,
+      ): void;
+      toContainReactComponentTimes<Type extends string | ComponentType<any>>(
+        type: Type,
+        times: number,
+        props?: Partial<PropsFor<Type>>,
+      ): void;
+      toProvideReactContext<Type>(
+        context: ReactContext<Type>,
+        value?: Type,
+      ): void;
+      toContainReactText(text: string): void;
+      toContainReactHtml(text: string): void;
+    }
+  }
+}
